% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/good_bye.R
\name{good_bye}
\alias{good_bye}
\title{good_bye}
\usage{
good_bye(name = "world")
}
\arguments{
\item{name}{a name; default is 'world'}
}
\value{
character a string; default is 'Hello, world!'
}
\description{
A one liner.
}
\details{
Here more details
}
\examples{
# Example usage of 'good_bye'

good_bye() 
}
\author{
author
}
\seealso{
Other example: \code{\link{hello}}
}

