% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{hello_world}
\usage{
hello_world()
}
\value{
character a string; default is 'Hello, world!'
}
\description{
The typical hello world example. Demonstrates also how to name functions and
the effect of family.
}
\examples{
# Example usage of hello()

hello()

hello("all") 
}
\author{
Reinhard Simon
}
\seealso{
Other example: \code{\link{hello}}
}

