% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io_dist_plot.R
\name{io_dist_plot}
\alias{io_dist_plot}
\title{io_dist_plot}
\usage{
io_dist_plot(input, output)
}
\arguments{
\item{input}{a server input}

\item{output}{a server output}
}
\value{
shiny plot element
}
\description{
Some shiny output
}

